requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==3.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.0
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
Django==4.2.15
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
firecrawl-py==0.0.20
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gcloud==0.18.3
google-auth==2.32.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
incremental==24.7.2
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
kubernetes==30.1.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.98
markdown-it-py==3.0.0
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
msgpack-python==0.5.6
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.40.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
posthog==3.5.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
PyYAML==6.0.1
redis==2.10.6
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rich==13.7.1
rsa==4.9
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
starlette==0.37.2
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
zope.interface==7.0.2
